requirements.txt
psutil>=5.9.0
requests>=2.31.0
google-generativeai>=0.3.2
python-dotenv>=1.0.0
readline>=6.2.4.1; sys_platform != "win32"

